@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: Unlicense OR MIT
 
+//go:build darwin && ios
 // +build darwin,ios
 
 package wm
@@ -26,13 +27,13 @@ import (
 )
 
 type context struct {
-	owner                    *window
-	c                        *gl.Functions
-	ctx                      C.CFTypeRef
-	layer                    C.CFTypeRef
-	init                     bool
-	frameBuffer              gl.Framebuffer
-	colorBuffer, depthBuffer gl.Renderbuffer
+	owner       *window
+	c           *gl.Functions
+	ctx         C.CFTypeRef
+	layer       C.CFTypeRef
+	init        bool
+	frameBuffer gl.Framebuffer
+	colorBuffer gl.Renderbuffer
 }
 
 func init() {
@@ -64,6 +65,10 @@ func contextAPI() gpu.OpenGL {
 	return gpu.OpenGL{}
 }
 
+func (c *context) RenderTarget() gpu.RenderTarget {
+	return gpu.OpenGLRenderTarget(c.frameBuffer)
+}
+
 func (c *context) API() gpu.API {
 	return contextAPI()
 }
@@ -75,7 +80,6 @@ func (c *context) Release() {
 	C.gio_renderbufferStorage(c.ctx, 0, C.GLenum(gl.RENDERBUFFER))
 	c.c.DeleteFramebuffer(c.frameBuffer)
 	c.c.DeleteRenderbuffer(c.colorBuffer)
-	c.c.DeleteRenderbuffer(c.depthBuffer)
 	C.gio_makeCurrent(0)
 	C.CFRelease(c.ctx)
 	c.ctx = 0
@@ -85,10 +89,6 @@ func (c *context) Present() error {
 	if c.layer == 0 {
 		panic("context is not active")
 	}
-	// Discard depth buffer as recommended in
-	// https://developer.apple.com/library/archive/documentation/3DDrawing/Conceptual/OpenGLES_ProgrammingGuide/WorkingwithEAGLContexts/WorkingwithEAGLContexts.html
-	c.c.BindFramebuffer(gl.FRAMEBUFFER, c.frameBuffer)
-	c.c.InvalidateFramebuffer(gl.FRAMEBUFFER, gl.DEPTH_ATTACHMENT)
 	c.c.BindRenderbuffer(gl.RENDERBUFFER, c.colorBuffer)
 	if C.gio_presentRenderbuffer(c.ctx, C.GLenum(gl.RENDERBUFFER)) == 0 {
 		return errors.New("presentRenderBuffer failed")
@@ -96,15 +96,18 @@ func (c *context) Present() error {
 	return nil
 }
 
-func (c *context) Lock() {}
-
-func (c *context) Unlock() {}
-
-func (c *context) Refresh() error {
+func (c *context) Lock() error {
+	if C.gio_makeCurrent(c.ctx) == 0 {
+		return errors.New("[EAGLContext setCurrentContext] failed")
+	}
 	return nil
 }
 
-func (c *context) MakeCurrent() error {
+func (c *context) Unlock() {
+	C.gio_makeCurrent(0)
+}
+
+func (c *context) Refresh() error {
 	if C.gio_makeCurrent(c.ctx) == 0 {
 		return errors.New("[EAGLContext setCurrentContext] failed")
 	}
@@ -112,7 +115,6 @@ func (c *context) MakeCurrent() error {
 		c.init = true
 		c.frameBuffer = c.c.CreateFramebuffer()
 		c.colorBuffer = c.c.CreateRenderbuffer()
-		c.depthBuffer = c.c.CreateRenderbuffer()
 	}
 	if !c.owner.isVisible() {
 		// Make sure any in-flight GL commands are complete.
@@ -124,14 +126,9 @@ func (c *context) MakeCurrent() error {
 	if C.gio_renderbufferStorage(c.ctx, c.layer, C.GLenum(gl.RENDERBUFFER)) == 0 {
 		return errors.New("renderbufferStorage failed")
 	}
-	w := c.c.GetRenderbufferParameteri(gl.RENDERBUFFER, gl.RENDERBUFFER_WIDTH)
-	h := c.c.GetRenderbufferParameteri(gl.RENDERBUFFER, gl.RENDERBUFFER_HEIGHT)
-	c.c.BindRenderbuffer(gl.RENDERBUFFER, c.depthBuffer)
-	c.c.RenderbufferStorage(gl.RENDERBUFFER, gl.DEPTH_COMPONENT16, w, h)
 	c.c.BindRenderbuffer(gl.RENDERBUFFER, currentRB)
 	c.c.BindFramebuffer(gl.FRAMEBUFFER, c.frameBuffer)
 	c.c.FramebufferRenderbuffer(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT0, gl.RENDERBUFFER, c.colorBuffer)
-	c.c.FramebufferRenderbuffer(gl.FRAMEBUFFER, gl.DEPTH_ATTACHMENT, gl.RENDERBUFFER, c.depthBuffer)
 	if st := c.c.CheckFramebufferStatus(gl.FRAMEBUFFER); st != gl.FRAMEBUFFER_COMPLETE {
 		return fmt.Errorf("framebuffer incomplete, status: %#x\n", st)
 	}
