@@ -59,19 +59,15 @@ type FrameEvent struct {
 type Callbacks interface {
 	SetDriver(d Driver)
 	Event(e event.Event)
-	// Func runs a function during an Event. This is required for platforms
-	// that require coordination between the rendering goroutine and the system
-	// main thread.
-	Run(f func())
 }
 
 type Context interface {
 	API() gpu.API
+	RenderTarget() gpu.RenderTarget
 	Present() error
-	MakeCurrent() error
 	Refresh() error
 	Release()
-	Lock()
+	Lock() error
 	Unlock()
 }
 
