@@ -5,8 +5,7 @@ package gpu
 import (
 	"fmt"
 
-	"github.com/cybriq/giocore/f32"
-	"github.com/cybriq/giocore/internal/ops"
+	"gioui.org/f32"
 )
 
 type resourceCache struct {
@@ -19,7 +18,7 @@ type resourceCache struct {
 // since benchmarking showed them as a bottleneck.
 type opCache struct {
 	// store the index + 1 in cache this key is stored in
-	index map[ops.Key]int
+	index map[opKey]int
 	// list of indexes in cache that are free and can be used
 	freelist []int
 	cache    []opCacheValue
@@ -27,12 +26,10 @@ type opCache struct {
 
 type opCacheValue struct {
 	data pathData
-	// computePath is the encoded path for compute.
-	computePath encoder
 
 	bounds f32.Rectangle
 	// the fields below are handled by opCache
-	key  ops.Key
+	key  opKey
 	keep bool
 }
 
@@ -73,7 +70,8 @@ func (r *resourceCache) frame() {
 }
 
 func (r *resourceCache) release() {
-	for _, v := range r.newRes {
+	r.frame()
+	for _, v := range r.res {
 		v.release()
 	}
 	r.newRes = nil
@@ -82,13 +80,13 @@ func (r *resourceCache) release() {
 
 func newOpCache() *opCache {
 	return &opCache{
-		index:    make(map[ops.Key]int),
+		index:    make(map[opKey]int),
 		freelist: make([]int, 0),
 		cache:    make([]opCacheValue, 0),
 	}
 }
 
-func (r *opCache) get(key ops.Key) (o opCacheValue, exist bool) {
+func (r *opCache) get(key opKey) (o opCacheValue, exist bool) {
 	v := r.index[key]
 	if v == 0 {
 		return
@@ -97,7 +95,7 @@ func (r *opCache) get(key ops.Key) (o opCacheValue, exist bool) {
 	return r.cache[v-1], true
 }
 
-func (r *opCache) put(key ops.Key, val opCacheValue) {
+func (r *opCache) put(key opKey, val opCacheValue) {
 	v := r.index[key]
 	val.keep = true
 	val.key = key
